refactor(modbus): relocate TCP TLS glue into tcp::tls for reuse
Moves ClientStream, build_client_tls_config, resolve_server_identity,
generate_self_signed, and build_server_tls_config out of tcp::client and
tcp::server into tcp::tls (now pub(crate)), so the upcoming rtu_over_tcp
transport can reuse the same TLS handling verbatim per MB-R-115 instead of
duplicating it. Pure move, no behavior change.
